+++ argopy/related/__init__.py
@@ -1,7 +1,9 @@
-from .topo import TopoFetcher
+from .topography import TopoFetcher
+from .ocean_ops_deployments import OceanOPSDeployments
 
 #
 __all__ = (
     # Classes:
     "TopoFetcher",
+    "OceanOPSDeployments",
 )
